+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 (2025-08-07)
+
+### Added
+
+- Implemented click cli
+- Made all code mypy strict compatible
+- Increase test coverage to >90%
+- Added unit tests
+  - dice test
+  - board test
+  - strategy test
+  - cli test
+- Added integration tests
+  - basic example game test
+  - cli module tests
+  - click cli tests
+
+### Removed
+
+- Some unused code.
+
 ## 0.2.0 (2025-08-06)
 
 ### Added